requirements.txt
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
click==7.1.1
itsdangerous==1.1.0
honeycomb-beeline<3,>=2.12.1
libhoney<2,>=1.9.0
google-cloud-bigquery<3,>=2.34.0
